refactor : 매칭 컴포넌트, 페이지 정리
